+++ fidget/src/core/types/mod.rs
@@ -1,0 +1,6 @@
+//! Custom types used during evaluation
+
+mod grad;
+mod interval;
+pub use grad::Grad;
+pub use interval::Interval;